fix(bashrc): fix uname for MobaXTerm
When starting, mobaxterm set an alias to uname, so the `uname -o` returns
Cygwin.  But this alias set seems to be later than the home user bashrc
parsing.  So `uname -o` in this context was returning 'GNU/Linux', which
is obviously not helpfull. `uname` still return CYGWIN_NT-10.0-WOW so
keeping it this way is simpler.